Delete orphaned ArgoCD Apps in Namespace Reconciler.
